+++ src/config/backpack/permissionmanager.php
@@ -1,0 +1,28 @@
+<?php
+
+return [
+
+    /*
+    | Backpack/PermissionManager configs.
+    */
+
+
+    /*
+    |--------------------------------------------------------------------------
+    | Disallow the user interface for creating/updating permissions or roles.
+    |--------------------------------------------------------------------------
+    | Roles and permissions are used in code by their name
+    | - ex: $user->hasPermissionTo('edit articles');
+    |
+    | So after the developer has entered all permissions and roles, the administrator should either:
+    | - not have access to the panels
+    | or
+    | - creating and updating should be disabled
+    */
+
+    'allow_permission_create' => true,
+    'allow_permission_update' => true,
+    'allow_role_create' => true,
+    'allow_role_update' => true,
+
+];